requirement.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-rest-authtoken==2.1.4
django-seed==0.3.1
djangorestframework==3.15.2
drf-yasg==1.21.8
Faker==30.8.2
inflection==0.5.1
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
